Creating game uses own controller.
