+++ app/angular/components/games/create-game/create-game-template.html
@@ -3,12 +3,12 @@
     <div class="gff-box col-xs-12 bg-white">
       <h2>Create a Game</h2>
       <p>Please enter game information below.</p>
-      <form class="new-game-form col-md-12" name="new-game" ng-submit="gc.createGame(gc.game)">
+      <form class="new-game-form col-md-12" name="new-game" ng-submit="cgc.createGame(cgc.game)">
         <div class="create-game-container">
           <div class="course-container">
             <p>Course Played</p>
             <div class="gff-select course">
-              <select ng-model="gc.game.name">
+              <select ng-model="cgc.game.name">
                 <option value="Newcastle GC - China Creek">Newcastle GC - China Creek</option>
                 <option value="Newcastle GC - Coal Creek">Newcastle GC - Coal Creek</option>
               </select>
@@ -17,32 +17,32 @@
 
           <div class="location-container">
             <p>City, State</p>
-            <input type="text" required ng-model="gc.game.location" placeholder="Seattle, WA"></input>
+            <input type="text" required ng-model="cgc.game.location" placeholder="Seattle, WA"></input>
           </div>
 
           <div class="date-container">
             <p>Date Played</p>
-            <input type="date" required ng-model="gc.game.datePlayed">
+            <input type="date" required ng-model="cgc.game.datePlayed">
           </div>
 
           <div class="players-container">
             <p>Players</p>
             <ul class="player-list">
-              <li class="player-box" ng-repeat="player in gc.players" ng-init="player.strokes=0">
+              <li class="player-box" ng-repeat="player in cgc.players" ng-init="player.strokes=0">
                 <div class="title-box">
                   <p>{{player.fullName}}</p>
                 </div>
-                <div class="strokes-box" ng-hide="gc.editing">
+                <div class="strokes-box" ng-hide="cgc.editing">
                   <p><strong><span class="color-green">+</span></strong>{{player.strokes}}</p>
                 </div>
-                <div class="edit-box" ng-hide="gc.editing">
-                  <a ng-click="gc.editing=true"<i class="fa fa-pencil"></i></a>
+                <div class="edit-box" ng-hide="cgc.editing">
+                  <a ng-click="cgc.editing=true"<i class="fa fa-pencil"></i></a>
                 </div>
-                <div class="edit-box" ng-show="gc.editing">
+                <div class="edit-box" ng-show="cgc.editing">
                   <input type="text" required ng-model="player.strokes" placeholder="{{player.strokes}}" maxlength="2"></input>
-                  <a ng-click="gc.editing=false"><i class="fa fa-floppy-o"></i></a>
+                  <a ng-click="cgc.editing=false"><i class="fa fa-floppy-o"></i></a>
                 </div>
-                <a class="remove-box" ng-click="gc.removeUser(player)">
+                <a class="remove-box" ng-click="cgc.removeUser(player)">
                   <i class="fa fa-minus-circle"></i>
                 </a>
               </li>
@@ -51,11 +51,11 @@
             <div class="add-players-container">
               <p>Add More Players</p>
               <ul class="player-list">
-                <li class="player-box" ng-repeat="friend in allFriends.friends">
+                <li class="player-box" ng-repeat="friend in cgc.allFriends.friends">
                   <div class="title-box">
                     <p>{{friend.fullName}}</p>
                   </div>
-                  <a class="add-box" ng-click="gc.addUser(friend)">
+                  <a class="add-box" ng-click="cgc.addUser(friend)">
                     <i class="fa fa-plus"></i>
                   </a>
                 </li>
