+++ app/angular/components/games/create-game/index.js
@@ -3,10 +3,10 @@
 module.exports = function(app) {
   app.component('createGame', {
     template: require('./create-game-template.html'),
-    controller: 'GamesController',
-    controllerAs: 'gc',
-    bindings: {
-      players: '=',
-    },
+    controller: 'CreateGameController',
+    controllerAs: 'cgc',
+    // bindings: {
+    //   players: '=',
+    // },
   });
 };
