+++ app/angular/components/games/games-template.html
@@ -8,4 +8,4 @@
   </div>
 </div>
 
-<create-game ng-show="gc.creating" data-players="gc.players"></create-game>
+<create-game ng-show="gc.creating"></create-game>
