--- app/angular/controllers/new_game_controller.js
@@ -1,13 +1,0 @@
-'use strict';
-
-module.exports = function(app) {
-  app.controller('NewGameController', ['$rootScope', function($rs) {
-
-    let ctrl = this;
-    ctrl.players = [];
-
-    ctrl.init = function() {
-      
-    }
-  }]);
-}
