+++ app/angular/scss/pages/base.scss
@@ -5,13 +5,13 @@ html, body, main {
 }
 
 html {
-  background: #ccc;
+  background: #fff;
 }
 
 main {
   display: table;
   z-index: 1;
-  background: #ccc;
+  background: #fff;
 }
 
 hr {
